#18 rename: infrastructure 패키지명 오타 수정
