+++ src/main/java/com/webeye/backend/explanation/application/ExplanationService.java
@@ -3,7 +3,7 @@ package com.webeye.backend.explanation.application;
 import com.webeye.backend.explanation.dto.response.DetailExplanationResponse;
 import com.webeye.backend.explanation.dto.response.PointExplanationResponse;
 import com.webeye.backend.imageanalysis.dto.request.ImageAnalysisRequest;
-import com.webeye.backend.imageanalysis.infrastrucutre.OpenAiClient;
+import com.webeye.backend.imageanalysis.infrastructure.OpenAiClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
